Attempt to fix by creating variable
Attemped to see if it would change by making it a variable.

Runs through to line 43 print statement for error testing, "Data outputted to json.dump" - but the file created just reads "null". So, it didn't work.

But it did work before attempting to write to json, so there is something going wrong with json code.